English grammar accessiblity
The first change mirrors Cypress's docs approach to English. The .it('method allows for code that looks like plain English')--this change is only a suggestion.

The second change eliminates a grammatical hiccup. Yes, technically "we can access to," is passable English. "We can gain access to," or "We can access," are what you'll hear native speakers say.

BTW, THANK YOU! I love this resource, and look forward to completing the course.